requirements.txt
Django>=4.1.1,<4.2
djangorestframework>=3.14,<3.15
psycopg2-binary>= 2.9.3,<2.10